chore: Replace browser-based MainViewIT with browserless MainViewTest
Convert integration test to browserless test using SpringBrowserlessTest,
removing the need for a running browser and server during testing.
